cap_genie_dist/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "pybind11>=2.10.0"]
build-backend = "setuptools.build_meta"

[project]
name = "capgenie"
version = "0.1.0"
description = "A bioinformatics tool for sequence analysis"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "pyahocorasick>=2.0.0",
    "biopython>=1.79",
    "plotly>=5.0.0",
    "pandas>=1.5.0",
    "pybind11>=2.10.0",
    "inquirer>=2.7.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
    "umap-learn>=0.5.0",
    "logomaker>=0.8",
    "scipy>=1.7.0",
    "numpy>=1.21.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

[tool.setuptools]
packages = ["capgenie"]
package-dir = {"" = "src"}

[project.scripts]
capgenie = "capgenie.cli:main"